# openapi.yaml
openapi: "3.0.3"
info:
  title: Weft API
  description: |
    Crowdsourced Civic Issue Reporting Platform.
    All endpoints are prefixed with `/v1`.
  version: "1.0.0"
  contact:
    name: Weft Engineering
    url: https://weft.city
servers:
  - url: http://localhost:8000/v1
    description: Local development
  - url: https://api.weft.city/v1
    description: Production
...
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      summary: Register a new citizen account
      operationId: registerCitizen
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegisterResponse"
        "409":
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      tags: [Auth]
      summary: Login with email + password
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: >-
            Login successful. Both tokens are returned in the body; no cookie
            is set.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/oauth/google:
    post:
      tags: [Auth]
      summary: Google OAuth sign-in
      operationId: googleOAuth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [id_token]
              properties:
                id_token:
                  type: string
                  description: Google ID token from mobile/web SDK
      responses:
        "200":
          description: >-
            OAuth successful. Both tokens are returned in the body; no cookie
            is set.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/oauth/apple:
    post:
      tags: [Auth]
      summary: Apple OAuth sign-in
      operationId: appleOAuth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [authorization_code]
              properties:
                authorization_code:
                  type: string
                name:
                  type: string
                  description: Full name (Apple provides only on first auth)
      responses:
        "200":
          description: >-
            OAuth successful. Both tokens are returned in the body; no cookie
            is set.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Rotate refresh token and get new access token
      description: >-
        Takes the refresh token from the JSON body. There is no cookie
        channel: no cookie is read, and none is set. Every client — Flutter and
        the React dashboard alike — holds the token itself and sends it here.

        The presented token is revoked in the same transaction that issues its
        replacement, so a refresh token is never valid twice. Both a new access
        token and a new refresh token are always returned: clients replace the
        stored refresh token on every rotation, so returning only one of them
        would silently end the session.

        Omitting `refresh_token` is a 400 (`MISSING_REFRESH_TOKEN`); an
        unknown, expired, already-rotated or revoked token is a 401, as is a
        token whose owner has been deactivated. The request body itself may be
        omitted entirely, which is the same 400 — the field is not declared
        `required` so that the missing-token case is a domain error rather than
        a schema-validation 422.
      operationId: refreshToken
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                refresh_token:
                  type: string
                  description: >-
                    Opaque refresh token issued by login or by a previous
                    rotation. Absent or empty is a 400.
      responses:
        "200":
          description: >-
            Token rotated. A new access token and a new refresh token are both
            returned in the body; no cookie is set.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "400":
          description: No refresh_token supplied in the body
        "401":
          description: Invalid, expired or revoked refresh token
  /auth/logout:
    post:
      tags: [Auth]
      summary: Revoke a refresh token (single device)
      description: >-
        Revokes one refresh token, ending the session on that device only. The
        token to revoke is supplied in the JSON body; there is no cookie
        channel, and none is set or cleared. The access token's `jti`
        identifies the access token, not the refresh token, so the token to
        revoke cannot be inferred and must be named explicitly.

        Deliberately unauthenticated — no `Authorization` header is required
        or inspected. Possession of the 64-byte refresh token is itself the
        proof, and demanding a live access token would make an expired session
        impossible to revoke, which is exactly when revocation matters.

        Idempotent by construction: an unknown token, an already-revoked
        token, an empty body and no body at all all return 204, so the
        response is never an oracle for which tokens exist.
      operationId: logout
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                refresh_token:
                  type: string
                  description: >-
                    Refresh token to revoke. Omitting it is not an error — the
                    call succeeds and revokes nothing.
      responses:
        "204":
          description: >-
            Token revoked, or no matching token existed. Indistinguishable by
            design.

  # ── Issues ──────────────────────────────────────────────────────────
...
  /users/me:
    get:
      tags: [Users]
      summary: Get own profile + trust score
      operationId: getMyProfile
      security:
        - BearerAuth: []
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"

    patch:
      tags: [Users]
      summary: Update profile (name, lang, FCM token)
      operationId: updateMyProfile
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  maxLength: 100
                preferred_lang:
                  type: string
                  enum: [en, hi, ta, te, mr, bn]
                fcm_token:
                  type: string
      responses:
        "200":
          description: Profile updated
...
  /admin/authority-users:
    post:
      tags: [Admin]
      summary: Create a new authority account
      operationId: createAuthorityUser
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAuthorityRequest"
      responses:
        "201":
          description: Authority account created
  /admin/authority-users/{user_id}/deactivate:
    patch:
      tags: [Admin]
      summary: Deactivate authority account
      operationId: deactivateAuthorityUser
      security:
        - BearerAuth: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Account deactivated
  /admin/departments:
    get:
      tags: [Admin]
      summary: List all departments
      operationId: listDepartments
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Department list

    post:
      tags: [Admin]
      summary: Create a new department
      operationId: createDepartment
      security:
        - BearerAuth: []
      responses:
        "201":
          description: Department created
  /admin/zones:
    post:
      tags: [Admin]
      summary: Create a new geographic zone
      operationId: createZone
      security:
        - BearerAuth: []
      responses:
        "201":
          description: Zone created
  /admin/system/stats:
    get:
      tags: [Admin]
      summary: System-wide usage statistics
      operationId: systemStats
      security:
        - BearerAuth: []
      responses:
        "200":
          description: System stats

# ═══════════════════════════════════════════════════════════════════════
# Components
# ═══════════════════════════════════════════════════════════════════════
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: RS256 signed JWT access token (24h expiry)
...
  schemas:
...
    UserRole:
      type: string
      enum: [CITIZEN, AUTHORITY, ADMIN]
...
    RegisterRequest:
      type: object
      required: [email, password, name]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        name:
          type: string
          maxLength: 100
        phone:
          type: string
    RegisterResponse:
      type: object
      required: [user_id, email, name, role, created_at]
      properties:
        user_id:
          type: string
          format: uuid
        email:
          type: string
        name:
          type: string
        role:
          $ref: "#/components/schemas/UserRole"
        created_at:
          type: string
          format: date-time
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    AuthResponse:
      type: object
      # `token_type` is omitted: it carries a default ("bearer") and so is not
      # required on the wire, exactly as the app's generated schema reports it.
      required: [access_token, refresh_token, expires_in, user]
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        token_type:
          type: string
          enum: [bearer]
        expires_in:
          type: integer
          description: Access token expiry in seconds
        user:
          $ref: "#/components/schemas/UserProfile"
    TokenResponse:
      type: object
      # Both tokens are always returned by /auth/refresh; a client that treated
      # refresh_token as optional would drop the rotated token and lose the
      # session on the next call.
      required: [access_token, refresh_token, expires_in]
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        token_type:
          type: string
        expires_in:
          type: integer

    # ── Issues ────────────────────────────────────────────────────────
...
    UserProfile:
      type: object
      # email, name and title are nullable for anonymous or phone-only accounts
      # and are therefore not required.
      required: [id, role, trust_score, total_points, is_anonymous, preferred_lang, created_at]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        name:
          type: string
        role:
          $ref: "#/components/schemas/UserRole"
        trust_score:
          type: number
        total_points:
          type: integer
        title:
          type: string
        is_anonymous:
          type: boolean
        preferred_lang:
          type: string
          # Same six languages the updateMyProfile body accepts. A client may
          # only ever send one of these, so it must only ever receive one.
          enum: [en, hi, ta, te, mr, bn]
        created_at:
          type: string
          format: date-time
...
    CreateAuthorityRequest:
      type: object
      required: [email, name, employee_id, department_id]
      properties:
        email:
          type: string
          format: email
        name:
          type: string
        password:
          type: string
          minLength: 8
        employee_id:
          type: string
        department_id:
          type: string
          format: uuid
        designation:
          type: string
        zone_ids:
          type: array
          items:
            type: string
            format: uuid

    # ── Common ────────────────────────────────────────────────────────
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          # `details` carries a default of {} and so is not required.
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
...